be: decouple netty and game
