+++ app/src/main/java/transport/config/GameServerInitializer.java
@@ -1,23 +1,13 @@
-package gameserver.config;
+package transport.config;
 
-import game.controller.Game;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelPipeline;
-import io.netty.channel.group.ChannelGroup;
 import io.netty.handler.codec.http.HttpObjectAggregator;
 import io.netty.handler.codec.http.HttpServerCodec;
 import io.netty.handler.codec.http.websocketx.WebSocketServerProtocolHandler;
 
 public class GameServerInitializer extends ChannelInitializer<Channel> {
-
-    private final ChannelGroup channelGroup;
-    private final Game game;
-
-    public GameServerInitializer(ChannelGroup group, Game game) {
-        this.channelGroup = group;
-        this.game = game;
-    }
 
     @Override
     protected void initChannel(Channel ch) throws Exception {
@@ -26,6 +16,6 @@ public class GameServerInitializer extends ChannelInitializer<Channel> {
                 .addLast(new HttpObjectAggregator(64 * 1024))
                 .addLast(new HttpRequestHandler("/ws"))
                 .addLast(new WebSocketServerProtocolHandler("/ws"))
-                .addLast(new TextWebSocketFrameHandler(channelGroup, game));
+                .addLast(new TextWebSocketFrameHandler());
     }
 }
