+++ app/src/main/java/transport/handler/JSONOutboundHandler.java
@@ -1,0 +1,15 @@
+package transport.handler;
+
+import game.model.GameRound;
+import io.netty.handler.codec.http.websocketx.TextWebSocketFrame;
+
+public class JSONOutboundHandler {
+
+    public void sendUpdate(GameRound round) {
+
+        String update = round.playersJSON().toString();
+
+        round.getChannelGroup().writeAndFlush(new TextWebSocketFrame(update));
+
+    }
+}
