+++ documents/project-outline/project-outline.tex
@@ -224,7 +224,7 @@
             \end{compactitem}\\
             \toprule
             \rowcolor{lightgray}
-            \textbf{Meilenstein 2} Woche: \textbf{2 Ende}\\
+            \textbf{Meilenstein 2} Woche: \textbf{6 Ende}\\
             \rowcolor{lightgray}
             \begin{compactitem}
                 \item \textbf{Alle Anforderungen stabilisiert.}
